requirements.txt
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
entsoe-py==0.5.8
Fiona==1.8.21
geojson-rewind==1.0.3
geopandas==0.11.1
idna==3.4
kaleido==0.2.1
munch==2.5.0
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
plotly==5.10.0
pyparsing==3.0.9
pyproj==3.4.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
Shapely==1.8.4
six==1.16.0
soupsieve==2.3.2.post1
tenacity==8.1.0
tweepy==4.10.1
urllib3==1.26.12
